rename upload folder to legacy_upload and test_upload_sender to test_legacy_upload_sender to prepare having new upload logic
